requirements.txt
supabase
python-dotenv
python-docx
langchain-groq
langchain-core
tenacity
pydantic-settings
fastapi
uvicorn